Push mutations that are only used in one place down the component tree as far as possible
